chore: combine streaming library, player and overlay
Monorepo
--------
Instead of keeping 3 separate repositories for related
packages, move the player and overlay library into this
repo and combine dependency management, build, lint, and
test tooling.

This simplifies upkeep of common dependencies as most
of the technology is shared. At the same time, changes
to a library are automatically available to the other
packages.

Git LFS
-------
To further reduce the impact of keeping deps in the repository, git LFS
is used to store the zip files externally while keeping the zero-install
advantages.

Additional changes:
-------------------
- replace jest with uvu
  (still an anomaly with the ws-source test that exits
  prematurely)
- implement more strict linting rules and apply them to
  the examples as well
- introduce justfile as a command runner